fix: audit wave 2026-09-03 - the remaining 60 findings
Architecture, security, test quality, typing and performance.

The registry is now the single source of truth for the code_audit package: the
package docstring catalogue, __all__ and the submodule-hiding tuple are all
derived from it, closing four findings that were one root cause. Thirteen
scanners had reimplemented "is this a test file" six incompatible ways; they
share one helper now. Provider pricing became one named tuple, ending accessors
that indexed [1] in one provider and [2] in another for the same value.

Security: Chrome keeps its renderer sandbox unless the process really is root,
so a renderer exploit from a scraped page no longer runs as the user holding the
DB and S3 credentials; proxy credentials are JSON-encoded into the auth
extension instead of interpolated raw; Authorization and Cookie headers are
dropped on a cross-host redirect; logged() routes parameters and exception text
through the redactor; and pip-audit runs weekly on one runner, which is the
security signal dependabot's disabled pip ecosystem was not providing.

Test quality: the two test-quality scanners that were never wired at tests/ now
run there - 60 vacuous assertions gained real ones and eight source-text
assertions became behavioural. The 1000-line module budget now covers tests/
too, so test_code_audit.py became tests/code_audit/, 89 files mirroring the
package it tests.

Typing: the batch-stamped ignore reason is gone from all 53 sites - 22 by
removing the cause, the rest by naming the real untyped source per site. Ruff's
version is now asserted against the declared pin rather than trusted, after the
audit ran under 0.15.9 while the repo declared 0.15.22.

Performance: vacuous_matching no longer calls ast.get_source_segment per
function before filtering, which was most of the tests-baseline scan.

+++ src/pyutilz/data/polarslib/binning.py
@@ -57,7 +57,7 @@ def _group_freqs(bins: pl.DataFrame, cols, drop_nulls: bool = False) -> Any:
 
 def _shannon_entropy(freqs: Any) -> float:
     """Shannon entropy (in nats) of a discrete frequency distribution that sums to 1."""
-    return -np.sum(freqs * np.log(freqs))  # type: ignore[no-any-return]  # untyped upstream source (json/external lib/dynamic attr); return value verified correct at runtime
+    return float(-np.sum(freqs * np.log(freqs)))  # float(): np.sum over an Any operand returns np.floating, not the declared float
 
 
 def entropy_for_column(bins: pl.DataFrame, col: str, drop_nulls: bool = False) -> float:
